+++ compiler/src/config.h
@@ -7,3 +7,5 @@
 #define ARGS_BUFFER_SIZE    512
 #define OUTPUT_BUFFER_SIZE  512
 #define SYM_TABLE_SIZE      512
+#define FILESTACK_DEPTH     5
+#define LINE_BUFFER_SIZE    120