+++ compiler/src/reader.c
@@ -8,10 +8,16 @@ int line_number = 1;
 
 FILE *fp;
 
+char filestack_position;
+FILE *filestack[FILESTACK_DEPTH];
+
 char poked = 0;
 char eof = 0;
 
 char buffer[ARGS_BUFFER_SIZE];
+
+char linepos;
+char linebuffer[LINE_BUFFER_SIZE];
 
 int buf_ptr;
 char is_buffered;
@@ -26,11 +32,50 @@ void init_reader(char *filename)
     line_number = 1;
     poked = 0;
     eof = 0;
+    filestack_position = 0;
+
+    fp = fopen(filename, "r");
+    filestack[filestack_position] = fp;
+
+    linepos = 0;
+    linebuffer[linepos] = 0;
+
+    if (!fp)
+        error(OPEN_FILE_ISSUE);
+}
+
+void shutdown_reader()
+{
+    for (int i = 0; i <= filestack_position; i++) {
+        if (filestack[i]) {
+            fclose(filestack[i]);
+        }
+    }
+}
+
+void include_file(char *filename)
+{
+    filestack_position++;
+
+    if (filestack_position >= FILESTACK_DEPTH) {
+        error(FILESTACK_OVERRUN);
+    }
 
     fp = fopen(filename, "r");
 
     if (!fp)
         error(OPEN_FILE_ISSUE);
+
+    filestack[filestack_position] = fp;
+}
+
+void filestack_pop()
+{
+    printf("File done!\r\n");
+    fclose(fp);
+    filestack_position--;
+
+    fp = filestack[filestack_position];
 }
 
 void enable_buffer()
@@ -157,11 +202,23 @@ char get_chr()
 
 
     if (!is_buffered) {
+      retry:
         if (!fread(&c, 1, 1, fp)) {
-            eof = 1;
+            if (filestack_position == 0) {
+                eof = 1;
+            } else {
+                filestack_pop();
+                goto retry;
+            }
         }
+
         if (buffer_enabled)
             buffer[buf_ptr] = c;
+
+        if (linepos < LINE_BUFFER_SIZE - 1) {
+            linebuffer[linepos++] = c;
+            linebuffer[linepos] = 0;
+        }
     } else {
         if (buffer_enabled)
             c = buffer[buf_ptr];
@@ -179,6 +236,8 @@ char get_chr()
 
 
     if (c == 10) {
+        linebuffer[0] = 0;
+        linepos = 0;
         line_number++;
     }